pyproject.toml
[project]
name = "piml-papers"
version = "0.1.0"
description = "Implementations of Physics Informed Machine Learning models"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "comet-ml==3.54.2",
    "flax==0.12.1",
    "hydra-core==1.3.2",
    "jax==0.8.1",
    "matplotlib==3.10.7",
    "numpy==2.3.5",
    "optax==0.2.6",
    "pillow>=12.0.0",
    "tensorboard==2.20.0",
    "torch==2.9.1",
    "tqdm==4.67.1",
    "wandb==0.23.0",
]

[dependency-groups]
dev = [
    "pre-commit==4.5.0",
    "ruff==0.14.6",
    "ty==0.0.1a28",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff]
line-length = 100
